Updates
Majority of changes related to plotting the results.  Now, whenever a person moves or his happiness status changes, the plot metadata for only that actor is updated.  The for loop controlling the iterations now allows for refreshing of the map plot each turn - since it should be faster and more efficient.  Also plotted each iteration is the happiness & movers plots.